+++ annotations/nocedit.exe/pseudocode/src/sound/sndmain.cpp/parseWavFile_FUN_005a3fe0.cpp
@@ -2,11 +2,11 @@
 // Address: 005a3fe0
 // Address Range: [[005a3fe0, 005a439a]]
 // Convention: __cdecl
-// Signature: int __cdecl sound_sndmain_cpp_parseWavFile_FUN_005a3fe0(_FILE *file_handle,long *file_offset_ptr,CSfxSample *sfx_sample)
+// Signature: int __cdecl sound_sndmain_cpp_parseWavFile_FUN_005a3fe0(_FILE *file_handle,int *file_offset_ptr,CSfxSample *sfx_sample)
 
 #include "nocturne.h"
 
-int __cdecl sound_sndmain_cpp_parseWavFile_FUN_005a3fe0(_FILE *file_handle,long *file_offset_ptr,CSfxSample *sfx_sample)
+int __cdecl sound_sndmain_cpp_parseWavFile_FUN_005a3fe0(_FILE *file_handle,int *file_offset_ptr,CSfxSample *sfx_sample)
 
 {
   short sVar1;
